Delete broken and unused State.set_value function
State.set_value was not used and was clearly misguided:

- it could only be called with a setting name that refers to an obs_source_t
- the caller would have to know to pass the name "text" as the property to actually update the referenced source's text data
- the function conflates the value parameter as both the contents of the OBS source text data, and the selected source name

It looks like the function could only have been called with the 'text_source' script preference as the value of source_name.  But this function would have both set that combo box to select the passed in value, as well updating the text of the selected source itself to be that value on screen in the OBS output video stream.