Add last_update note
* Factor out set_last_update_timestamp() function
* Plumb through a descriptive reason for the update timestamp
* Add a timestamp to the null combo box selection

The last_update timestamp in combinatino with the timestamp in the combo box clearly demonstrates that the "Reload Text Source list" button does induce a widget redraw, but that the combo box property list itself is not being updated.

The combo box update is probably going to a dangling reference, potentially causing memory corruption.  Saving obs_property_t references for later use is apparently not valid.