Fix "Reload Text Source list" button
Change fill_sources_property_list to take the setting name of the property list to update instead of a stale reference to a property object.

It turns out that references to property objects are not stable.  Fetch the correct property to update by name out of the full obs_properties_t list that OBS passed through the callback instead.

This finally gets the reload button working again.