Remove prop_ref from Script.properties
Storing references to properties is not a valid thing to do.

These references get invalidated and stale arbitrarily.  Perhaps due to reallocation during properties management?  For example, prior to the previous commit, having the "Reset Timer" button come after the Text Source list caused the Text Source property reference to change, but only when the Reset Timer button was part of the preferences list - not when added manually in script_properties().