Switch text_source list refresh back to references
Indeed, the bug experienced earlier with the text_source combo list not being updated via the dedicated refresh button was not actually caused by obs_property_t references not being stable.

Instead, it was caused by the button callback lambda closing over a reference to the current _prop variable used in the script_properties generation loop.  That means the button's callback was trying to update a combo list on an unrelated property - whichever was the last one generated via the declarative preferences mechanism.

Now that we have a separate add_combo_list_regeneration_button() function to capture the property reference in a dedicated Python variable for each combo list, the lambda callback is correctly able to target the correct property to do the list refresh on.

The main takeaway here is to remember that variables in Python loops have function scope, so captured references to them will always get their last assigned value (usually the one from the last iteration of the loop).