Fix text input jankiness
Change global_property_modification_handler to return False.  Returning True was being used to update the property widgets so the status could be displayed in an info property.

The problem is that when OBS redraws the widgets, it resets the cursor position to the end of the field the user is editing, which renders anything but the end of the field unusable.